Conflicts:
	.gradle/2.2.1/taskArtifacts/cache.properties.lock
	.gradle/2.2.1/taskArtifacts/fileHashes.bin
	.gradle/2.2.1/taskArtifacts/fileSnapshots.bin
	.gradle/2.2.1/taskArtifacts/outputFileStates.bin
	.gradle/2.2.1/taskArtifacts/taskArtifacts.bin
	.idea/libraries/appcompat_v7_21_0_3.xml
	.idea/libraries/play_services_6_5_87.xml
	.idea/libraries/play_services_base_6_5_87.xml
	.idea/libraries/play_services_wearable_6_5_87.xml
	.idea/libraries/recyclerview_v7_21_0_0.xml
	.idea/libraries/support_annotations_21_0_0.xml
	.idea/libraries/support_annotations_21_0_3.xml
	.idea/libraries/support_v4_21_0_0.xml
	.idea/libraries/support_v4_21_0_3.xml
	.idea/libraries/wearable_1_1_0.xml
	.idea/workspace.xml
	build/intermediates/model_data.bin
	local.properties
	sketchagram.iml

+++ mobile/src/main/java/sketchagram/chalmers/com/model/Painting.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 public class Painting extends AMessage {
 
-    protected Painting(double timestamp, ADigitalPerson sender, List<ADigitalPerson> receiver ) {
+    protected Painting(long timestamp, ADigitalPerson sender, List<ADigitalPerson> receiver ) {
         super(timestamp, sender, receiver);
     }
